requirements.txt
Flask
Flask-Cors
openai
deep-translator
gTTS
python-dotenv
